+++ fuzzing_cli/fuzz/config/options.py
@@ -1,7 +1,7 @@
 import base64
 import math
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import Any, Dict, List, Optional, Set, Tuple, Union
 
 import click
 
@@ -32,6 +32,9 @@ class FuzzingOptions:
         time_limit: Optional[str] = None,
         chain_id: Optional[Union[str, int]] = None,
         enable_cheat_codes: Optional[bool] = None,
+        foundry_tests: bool = False,
+        target_contracts: Optional[Dict[str, Set[str]]] = None,
+        _validate_key: bool = True,
     ):
         self.ide: Optional[str] = ide and ide.lower()
         self.quick_check = quick_check
@@ -56,12 +59,21 @@ class FuzzingOptions:
         self.enable_cheat_codes = (
             bool(enable_cheat_codes) if enable_cheat_codes is not None else None
         )
+        self.foundry_tests = foundry_tests
+        self.target_contracts = target_contracts
 
         self.auth_endpoint = None
         self.refresh_token = None
         self.auth_client_id = None
 
-        self.validate(key)
+        self.validate(key, _validate_key)
+
+        if key:
+            (
+                self.auth_endpoint,
+                self.auth_client_id,
+                self.refresh_token,
+            ) = self._decode_refresh_token(key)
 
         if type(additional_contracts_addresses) == str:
             self.additional_contracts_addresses: Optional[List[str]] = [
@@ -69,10 +81,6 @@ class FuzzingOptions:
             ]
         else:
             self.additional_contracts_addresses = additional_contracts_addresses
-
-        self.auth_endpoint, self.auth_client_id, self.refresh_token = self._decode_refresh_token(
-            key
-        )
 
     @staticmethod
     def parse_chain_id(chain_id: Optional[Union[str, int]]) -> Optional[str]:
@@ -113,7 +121,7 @@ class FuzzingOptions:
         ide: Optional[str] = None,
         deployed_contract_address: Optional[str] = None,
         targets: Optional[List[str]] = None,
-        map_to_original_source: bool = False,
+        map_to_original_source: Optional[bool] = None,
         corpus_target: Optional[str] = None,
         additional_contracts_addresses: Optional[Union[List[str], str]] = None,
         dry_run: bool = False,
@@ -124,6 +132,9 @@ class FuzzingOptions:
         build_directory: Optional[str] = None,
         sources_directory: Optional[str] = None,
         enable_cheat_codes: Optional[bool] = None,
+        foundry_tests: bool = False,
+        target_contracts: Optional[Dict[str, Set[str]]] = None,
+        _validate_key: bool = True,
     ) -> "FuzzingOptions":
         return cls.parse_obj(
             {
@@ -141,7 +152,11 @@ class FuzzingOptions:
                         "deployed_contract_address": deployed_contract_address
                         or config.get("deployed_contract_address"),
                         "targets": targets or config.get("targets"),
-                        "map_to_original_source": map_to_original_source,
+                        "map_to_original_source": config.get(
+                            "map_to_original_source", False
+                        )
+                        if map_to_original_source is None
+                        else map_to_original_source,
                         "rpc_url": config.get("rpc_url"),
                         "faas_url": config.get("faas_url"),
                         "number_of_cores": config.get("number_of_cores"),
@@ -150,10 +165,7 @@ class FuzzingOptions:
                         "additional_contracts_addresses": additional_contracts_addresses
                         or config.get("additional_contracts_addresses"),
                         "dry_run": dry_run,
-                        "key": key
-                        or config.get("key")
-                        or config.get("api_key")
-                        or config.get("refresh_token"),
+                        "key": key,
                         "project": project or config.get("project"),
                         "truffle_executable_path": truffle_executable_path,
                         "incremental": config.get("incremental"),
@@ -166,6 +178,9 @@ class FuzzingOptions:
                         )
                         if enable_cheat_codes is None
                         else enable_cheat_codes,
+                        "foundry_tests": foundry_tests,
+                        "target_contracts": target_contracts,
+                        "_validate_key": _validate_key,
                     }
                 ).items()
                 if v is not None
@@ -194,7 +209,7 @@ class FuzzingOptions:
             raise click.exceptions.UsageError(error_message)
         return endpoint, client_id, rt
 
-    def validate(self, key: Optional[str] = None):
+    def validate(self, key: Optional[str] = None, __validate_key: bool = True):
         if not self.build_directory:
             raise click.exceptions.UsageError(
                 "Build directory not provided. You need to set the `build_directory` "
@@ -206,10 +221,10 @@ class FuzzingOptions:
                 "please set the `sources_directory` under the `fuzz` key of your .fuzz.yml config file."
             )
 
-        if not key:
+        if not key and __validate_key is True:
             raise click.exceptions.UsageError(
                 "API key was not provided. You need to provide an API key as the `--key` parameter "
-                "of the `fuzz run` command or set the `key` under the `fuzz` key of your .fuzz.yml config file."
+                "of the `fuzz run` command or as `FUZZ_API_KEY` environment variable."
             )
         if not self.quick_check and not self.deployed_contract_address:
             raise click.exceptions.UsageError(