+++ tests/test_fuzz_lessons.py
@@ -46,7 +46,7 @@ def test_start(tmp_path: Path, hardhat_fuzzing_lessons_project, description: str
     }
 
 
-def test_already_started(tmp_path: Path, hardhat_fuzzing_lessons_project):
+def test_already_started(api_key, tmp_path: Path, hardhat_fuzzing_lessons_project):
     write_config(tmp_path)
 
     with mocked_rpc_client(
@@ -62,7 +62,7 @@ def test_already_started(tmp_path: Path, hardhat_fuzzing_lessons_project):
     assert tmp_path.joinpath(".fuzzing_lessons.json").exists() is True
 
 
-def test_stop(tmp_path, hardhat_fuzzing_lessons_project):
+def test_stop(api_key, tmp_path, hardhat_fuzzing_lessons_project):
     write_config(tmp_path)
 
     with mocked_rpc_client(
@@ -123,7 +123,9 @@ def test_stop(tmp_path, hardhat_fuzzing_lessons_project):
     }
 
 
-def test_stop_no_transactions_in_lesson(tmp_path, hardhat_fuzzing_lessons_project):
+def test_stop_no_transactions_in_lesson(
+    api_key, tmp_path, hardhat_fuzzing_lessons_project
+):
     write_config(tmp_path)
 
     with mocked_rpc_client(
@@ -150,7 +152,9 @@ def test_stop_no_transactions_in_lesson(tmp_path, hardhat_fuzzing_lessons_projec
 
 
 @pytest.mark.parametrize("command", ["stop", "abort"])
-def test_not_started(tmp_path: Path, hardhat_fuzzing_lessons_project, command: str):
+def test_not_started(
+    api_key, tmp_path: Path, hardhat_fuzzing_lessons_project, command: str
+):
     write_config(tmp_path)
 
     runner = CliRunner()
@@ -160,7 +164,7 @@ def test_not_started(tmp_path: Path, hardhat_fuzzing_lessons_project, command: s
     assert result.output == "Error: No fuzzing lesson is running\n"
 
 
-def test_abort(tmp_path, hardhat_fuzzing_lessons_project):
+def test_abort(api_key, tmp_path, hardhat_fuzzing_lessons_project):
     write_config(tmp_path)
 
     with mocked_rpc_client(