+++ HISTORY.rst
@@ -2,6 +2,24 @@
 History
 =======
 
+0.11.0 (2023-04-07)
+--------------------
+- Make :code:`no-assert` mode default for :code:`fuzz arm` command
+- Drop support for Python 3.6
+- Update dependencies
+- Fix various bugs
+- Improve error messages
+- Remove :code:`key` parameter from fuzzing config. Only :code:`FUZZ_API_KEY` environment variable or :code:`--key` command line argument is supported now.
+- Finalize Foundry seamless integration
+
+0.10.2 (2023-02-13)
+--------------------
+- Fix quickcheck campaigns bug
+
+0.10.1 (2023-02-13)
+--------------------
+- Add block data to transactions in corpus
+
 0.10.0 (2023-02-13)
 --------------------
 - Add foundry unit tests submission command (:code:`fuzz forge test`)